third_party/blink/web_tests/external/wpt/css/css-values/calc-sibling-function-expected.txt

This is a testharness.js-based test.
[FAIL] sibling-count on pseudo-element
  assert_equals: expected "4" but got "2"
Harness: the test ran to completion.
